Move existing django tests (#997)
