+++ tests/contrib/django_old/app/templates/cached_list.html
