+++ tests/contrib/django_old/compat.py
