+++ tests/contrib/django_old/conftest.py
@@ -4,7 +4,7 @@ from django.conf import settings
 
 # We manually designate which settings we will be using in an environment variable
 # This is similar to what occurs in the `manage.py`
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'tests.contrib.django.app.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'tests.contrib.django_old.app.settings')
 
 
 # `pytest` automatically calls this function once when tests are run.
