+++ tests/contrib/django_old/test_autopatching.py
