+++ tests/contrib/django_old/test_middleware.py
@@ -40,7 +40,7 @@ class DjangoMiddlewareTest(DjangoTraceTestCase):
         assert sp_request.get_tag('django.user.is_authenticated') == 'False'
         assert sp_request.get_tag('http.method') == 'GET'
         assert sp_request.span_type == 'http'
-        assert sp_request.resource == 'tests.contrib.django.app.views.UserList'
+        assert sp_request.resource == 'tests.contrib.django_old.app.views.UserList'
 
     def test_middleware_trace_request_qs(self):
         return self.test_middleware_trace_request('foo=bar')
@@ -166,7 +166,7 @@ class DjangoMiddlewareTest(DjangoTraceTestCase):
         span = spans[0]
         assert span.get_tag('http.status_code') == '403'
         assert span.get_tag(http.URL) == 'http://testserver/fail-view/'
-        assert span.resource == 'tests.contrib.django.app.views.ForbiddenView'
+        assert span.resource == 'tests.contrib.django_old.app.views.ForbiddenView'
 
     def test_middleware_trace_function_based_view(self):
         # ensures that the internals are properly traced when using a function views
@@ -180,7 +180,7 @@ class DjangoMiddlewareTest(DjangoTraceTestCase):
         span = spans[0]
         assert span.get_tag('http.status_code') == '200'
         assert span.get_tag(http.URL) == 'http://testserver/fn-view/'
-        assert span.resource == 'tests.contrib.django.app.views.function_view'
+        assert span.resource == 'tests.contrib.django_old.app.views.function_view'
 
     def test_middleware_trace_error_500(self):
         # ensures we trace exceptions generated by views
@@ -195,7 +195,7 @@ class DjangoMiddlewareTest(DjangoTraceTestCase):
         assert span.error == 1
         assert span.get_tag('http.status_code') == '500'
         assert span.get_tag(http.URL) == 'http://testserver/error-500/'
-        assert span.resource == 'tests.contrib.django.app.views.error_500'
+        assert span.resource == 'tests.contrib.django_old.app.views.error_500'
         assert 'Error 500' in span.get_tag('error.stack')
 
     def test_middleware_trace_callable_view(self):
@@ -210,7 +210,7 @@ class DjangoMiddlewareTest(DjangoTraceTestCase):
         span = spans[0]
         assert span.get_tag('http.status_code') == '200'
         assert span.get_tag(http.URL) == 'http://testserver/feed-view/'
-        assert span.resource == 'tests.contrib.django.app.views.FeedView'
+        assert span.resource == 'tests.contrib.django_old.app.views.FeedView'
 
     def test_middleware_trace_partial_based_view(self):
         # ensures that the internals are properly traced when using a function views
@@ -238,7 +238,7 @@ class DjangoMiddlewareTest(DjangoTraceTestCase):
         span = spans[0]
         assert span.get_tag('http.status_code') == '200'
         assert span.get_tag(http.URL) == 'http://testserver/lambda-view/'
-        assert span.resource == 'tests.contrib.django.app.views.<lambda>'
+        assert span.resource == 'tests.contrib.django_old.app.views.<lambda>'
 
     @modify_settings(
         MIDDLEWARE={
@@ -307,10 +307,10 @@ class DjangoMiddlewareTest(DjangoTraceTestCase):
 
     @modify_settings(
         MIDDLEWARE={
-            'append': 'tests.contrib.django.app.middlewares.HandleErrorMiddlewareSuccess',
+            'append': 'tests.contrib.django_old.app.middlewares.HandleErrorMiddlewareSuccess',
         },
         MIDDLEWARE_CLASSES={
-            'append': 'tests.contrib.django.app.middlewares.HandleErrorMiddlewareSuccess',
+            'append': 'tests.contrib.django_old.app.middlewares.HandleErrorMiddlewareSuccess',
         },
     )
     def test_middleware_handled_view_exception_success(self):
@@ -333,10 +333,10 @@ class DjangoMiddlewareTest(DjangoTraceTestCase):
 
     @modify_settings(
         MIDDLEWARE={
-            'append': 'tests.contrib.django.app.middlewares.HandleErrorMiddlewareClientError',
+            'append': 'tests.contrib.django_old.app.middlewares.HandleErrorMiddlewareClientError',
         },
         MIDDLEWARE_CLASSES={
-            'append': 'tests.contrib.django.app.middlewares.HandleErrorMiddlewareClientError',
+            'append': 'tests.contrib.django_old.app.middlewares.HandleErrorMiddlewareClientError',
         },
     )
     def test_middleware_handled_view_exception_client_error(self):
