+++ tests/contrib/djangorestframework_old/app/views.py
