+++ tox.ini
@@ -56,10 +56,10 @@ envlist =
 # Python 3.7 needs Celery 4.3
     celery_contrib-{py27,py34,py35,py36,py37}-celery43-redis320-kombu44
     dbapi_contrib-{py27,py34,py35,py36}
-    django_contrib{,_autopatch}-{py27,py34,py35,py36}-django{18,111}-djangopylibmc06-djangoredis45-pylibmc-redis{210}-memcached
-    django_contrib{,_autopatch}-{py34,py35,py36}-django{200}-djangopylibmc06-djangoredis45-pylibmc-redis{210}-memcached
-    django_drf_contrib-{py27,py34,py35,py36}-django{111}-djangorestframework{34,37,38}
-    django_drf_contrib-{py34,py35,py36}-django{200}-djangorestframework{37,38}
+    django_contrib_old{,_autopatch}-{py27,py34,py35,py36}-django{18,111}-djangopylibmc06-djangoredis45-pylibmc-redis{210}-memcached
+    django_contrib_old{,_autopatch}-{py34,py35,py36}-django{200}-djangopylibmc06-djangoredis45-pylibmc-redis{210}-memcached
+    django_drf_contrib_old-{py27,py34,py35,py36}-django{111}-djangorestframework{34,37,38}
+    django_drf_contrib_old-{py34,py35,py36}-django{200}-djangorestframework{37,38}
     elasticsearch_contrib-{py27,py34,py35,py36}-elasticsearch{16,17,18,23,24,51,52,53,54,63}
     elasticsearch_contrib-{py27,py34,py35,py36}-elasticsearch1{100}
     elasticsearch_contrib-{py27,py34,py35,py36}-elasticsearch2{50}
@@ -353,9 +353,9 @@ commands =
     cassandra_contrib: pytest {posargs} tests/contrib/cassandra
     celery_contrib: pytest {posargs} tests/contrib/celery
     dbapi_contrib: pytest {posargs} tests/contrib/dbapi
-    django_contrib: pytest {posargs} tests/contrib/django
-    django_contrib_autopatch: python tests/ddtrace_run.py pytest {posargs} tests/contrib/django
-    django_drf_contrib: pytest {posargs} tests/contrib/djangorestframework
+    django_contrib_old: pytest {posargs} tests/contrib/django_old
+    django_contrib_old_autopatch: python tests/ddtrace_run.py pytest {posargs} tests/contrib/django_old
+    django_drf_contrib_old: pytest {posargs} tests/contrib/djangorestframework_old
     elasticsearch_contrib: pytest {posargs} tests/contrib/elasticsearch
     falcon_contrib: pytest {posargs} tests/contrib/falcon/test_middleware.py tests/contrib/falcon/test_distributed_tracing.py
     falcon_contrib_autopatch: python tests/ddtrace_run.py pytest {posargs} tests/contrib/falcon/test_autopatch.py
